follow relay convention
