+++ python/tvm/relax/op/nn/__init__.py
@@ -14,11 +14,6 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-# pylint: disable=wildcard-import, redefined-builtin
-"""Relax core operators."""
-
-# Operators
-from .base import *
-from .tensor import *
-from .op_attrs import *
+# pylint: disable=wildcard-import
+"""Neural network related operators."""
 from .nn import *
