+++ python/tvm/relax/op/nn/_make.py
@@ -14,11 +14,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-# pylint: disable=wildcard-import, redefined-builtin
-"""Relax core operators."""
+"""Constructor APIs"""
+import tvm._ffi
 
-# Operators
-from .base import *
-from .tensor import *
-from .op_attrs import *
-from .nn import *
+tvm._ffi._init_api("relax.op.nn", __name__)
