+++ src/relax/op/tensor/binary.cc
@@ -26,6 +26,83 @@
 
 namespace tvm {
 namespace relax {
+
+Optional<Expr> InferShapeBinaryBroadcast(const Call& call, DiagnosticContext diag_ctx) {
+  if (call->args.size() != 2) {
+    diag_ctx.EmitFatal(Diagnostic::Error(call->span)
+                       << "Binary broadcast op should have 2 arguments");
+  }
+  Expr lhs_shape = call->args[0]->shape();
+  Expr rhs_shape = call->args[1]->shape();
+  auto* s0 = lhs_shape.as<ShapeExprNode>();
+  auto* s1 = rhs_shape.as<ShapeExprNode>();
+  if (s0 && s1) {
+    std::vector<PrimExpr> output_shape;
+    size_t ndim0 = s0->values.size();
+    size_t ndim1 = s1->values.size();
+    size_t i = 1;
+    for (; i <= std::min(ndim0, ndim1); ++i) {
+      PrimExpr dim0 = s0->values[ndim0 - i];
+      PrimExpr dim1 = s1->values[ndim1 - i];
+      if (EqualConstInt(dim0, 1)) {
+        output_shape.push_back(dim1);
+      } else if (EqualConstInt(dim1, 1)) {
+        output_shape.push_back(dim0);
+      } else if (EqualCheck(dim0, dim1)) {
+        output_shape.push_back(dim0);
+      } else {
+        // defer the computation of output shapes to runtime
+        // e.g., broadcast Tensor([m, n]), Tensor([k]) -> defer to runtime
+        return Call(ExternFunc(String("vm.binary_broadcast_shape_infer")),
+                    {call->args[0], call->args[1]}, {}, {});
+      }
+    }
+    size_t max_ndim = std::max(ndim0, ndim1);
+    auto& longer_shape = (ndim0 > ndim1) ? s0 : s1;
+    for (; i <= max_ndim; ++i) {
+      output_shape.push_back(longer_shape->values[max_ndim - i]);
+    }
+    return ShapeExpr(Array<PrimExpr>(output_shape.rbegin(), output_shape.rend()));
+  } else {
+    return NullOpt;
+  }
+}
+
+Type InferTypeBinaryBroadcast(const Call& call, DiagnosticContext diag_ctx) {
+  if (call->args.size() != 2) {
+    diag_ctx.EmitFatal(Diagnostic::Error(call->span)
+                       << "Binary broadcast op should have 2 arguments");
+  }
+  Type lhs_type = call->args[0]->checked_type();
+  Type rhs_type = call->args[1]->checked_type();
+  auto* t0 = lhs_type.as<DynTensorTypeNode>();
+  auto* t1 = rhs_type.as<DynTensorTypeNode>();
+  if (!t0 || !t1) {
+    diag_ctx.EmitFatal(Diagnostic::Error(call->span)
+                       << "Both lhs and rhs should be DynTensor for broadcasting, but got "
+                       << lhs_type->GetTypeKey() << " and " << rhs_type->GetTypeKey());
+  }
+
+  DataType output_dtype;
+  if (t0->IsUnknownDtype() || t1->IsUnknownDtype()) {
+    output_dtype = DataType::Void();
+  } else if (t0->dtype != t1->dtype) {
+    diag_ctx.EmitFatal(Diagnostic::Error(call->span)
+                       << "Data types " << t0->dtype << " and " << t1->dtype
+                       << " must be equal for broadcasting operators");
+  } else {
+    output_dtype = t0->dtype;
+  }
+
+  int output_ndim;
+  if (t0->IsUnknownNdim() || t1->IsUnknownNdim()) {
+    output_ndim = -1;
+  } else {
+    output_ndim = std::max(t0->ndim, t1->ndim);
+  }
+  return DynTensorType(output_ndim, output_dtype);
+}
+
 TVM_REGISTER_NODE_TYPE(Conv2dAttrs);
 
 RELAX_REGISTER_BINARY_BROADCAST_OP("add")
@@ -35,41 +112,6 @@ RELAX_REGISTER_BINARY_BROADCAST_OP("add")
 RELAX_REGISTER_BINARY_BROADCAST_OP("multiply")
     .describe("Elementwise multiply with broadcasting")
     .set_support_level(1);
-
-RELAY_REGISTER_OP("relax.dense")
-    .set_num_inputs(2)
-    .add_argument("e1", "Expr", "The input expression")
-    .add_argument("e2", "Expr", "The input expression")
-    .set_attr<FInferShape>("FInferShape", InferShapeDense)
-    .set_attr<FInferType>("FInferType", InferTypeDense);
-
-Expr MakeDense(Expr expr1, Expr expr2) {
-  static const Op& op = Op::Get("relax.dense");
-  return Call(op, {expr1, expr2}, {}, {});
-}
-
-TVM_REGISTER_GLOBAL("relax.op.dense").set_body_typed(MakeDense);
-
-RELAY_REGISTER_OP("relax.conv2d")
-    .set_num_inputs(2)
-    .add_argument("e1", "Expr", "The input expression")
-    .add_argument("e2", "Expr", "The input expression")
-    .set_attrs_type<Conv2dAttrs>()
-    .set_attr<FInferShape>("FInferShape", InferShapeConv2d)
-    .set_attr<FInferType>("FInferType", InferTypeDense);
-
-Expr MakeConv2d(Expr expr1, Expr expr2, Array<PrimExpr> kernel_size, Array<PrimExpr> stride,
-                Array<PrimExpr> padding, Array<PrimExpr> dilation) {
-  static const Op& op = Op::Get("relax.conv2d");
-  auto attrs = make_object<Conv2dAttrs>();
-  attrs->kernel_size = kernel_size;
-  attrs->stride = stride;
-  attrs->padding = padding;
-  attrs->dilation = dilation;
-  return Call(op, {expr1, expr2}, Attrs(attrs), {});
-}
-
-TVM_REGISTER_GLOBAL("relax.op.conv2d").set_body_typed(MakeConv2d);
 
 }  // namespace relax
 }  // namespace tvm
